+++ packages/bundle/src/index.js
@@ -1,0 +1,5 @@
+import create from 'web-speech-cognitive-services/lib/SpeechServices';
+
+export {
+  create
+}